coordinates.py = added a method to return a copy of the current Coordinates. manager.py = option for the tracking and the landmark process. coordswriter.py = modified the process and add the usable outputs videos. test_coordinates.py = added some tests. test_coordswriter.py = added some tests. personsbuffer.py = added methods to know if the tracking process has been used and same for the landmark process.

+++ sppas/src/imagedata/facelandmark.py
@@ -180,38 +180,38 @@ class FaceLandmark(object):
         #     self.__landmarks.append((x, y))
 
         for i in range(10):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 2
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 2)
             self.__landmarks.append((x, y))
 
         for i in range(10):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 3
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 3)
             self.__landmarks.append((x, y))
 
         for i in range(10):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 4
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 4)
             self.__landmarks.append((x, y))
 
         for i in range(10):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 5
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 5)
             self.__landmarks.append((x, y))
 
         for i in range(10):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 6
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 6)
             self.__landmarks.append((x, y))
 
         for i in range(10):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 7
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 7)
             self.__landmarks.append((x, y))
 
         for i in range(8):
-            x = 3*w/4 * ((i+1)/10)
-            y = (h/10) * 8
+            x = int(3*w/4 * ((i+1)/10))
+            y = int((h/10) * 8)
             self.__landmarks.append((x, y))
 
     # -----------------------------------------------------------------------
